requirements.txt
PyMuPDF>=1.24.0
Pillow>=10.0.0
numpy>=1.20.0
